Text Summarizer/Flask_work/requirements.txt
Flask==2.1.3
mysql_connector_repackaged==0.3.1
spacy==3.5.2
